% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{add_date_vector_2012}
\alias{add_date_vector_2012}
\title{Add the date vector to a simulation output table}
\usage{
add_date_vector_2012(sim_i, date)
}
\arguments{
\item{sim_i}{Simulation output table element i in the list of simulation outputs.}

\item{date}{Date vector which should be added.}
}
\description{
Add the date vector to a simulation output table
}
\keyword{internal}
