% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{add_id}
\alias{add_id}
\title{Add id column to output table}
\usage{
add_id(tbl, col_unique = "unit")
}
\arguments{
\item{tbl}{Output table read from the SWAT model results}

\item{col_unique}{Column names which should be considered to identify unique
combinations}
}
\description{
Add id column to output table
}
\keyword{internal}
