% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{add_obj}
\alias{add_obj}
\title{Add the unit values for which objects the parameter change should be applied}
\usage{
add_obj(par_cal, unit, unit_all)
}
\arguments{
\item{par_cal}{The calibration.cal tibble for the parameter i}

\item{unit}{Character string that defines condition for the object units (ids of e.g. hru objects)}

\item{unit_all}{List of vectors that define the unit ids of all object types.}
}
\description{
Add the unit values for which objects the parameter change should be applied
}
\keyword{internal}
