% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{add_quotes_if_chr}
\alias{add_quotes_if_chr}
\title{Add quotes to character values of vector in filter expression}
\usage{
add_quotes_if_chr(chr)
}
\arguments{
\item{chr}{Character string providing values for filter variable}
}
\description{
Add quotes to character values of vector in filter expression
}
\keyword{internal}

