% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{add_run_info}
\alias{add_run_info}
\title{Prepare run info for the simulation experiment}
\usage{
add_run_info(run_info, sim_result, run_index, t1)
}
\arguments{
\item{run_index}{Vector of run IDs}

\item{t1}{End date time stamp}

\item{sim_results}{List of simulation results from the SWAT model runs}

\item{output}{Table of defined output variables}

\item{project_path}{Path to the SWAT project folder}
}
\description{
Prepare run info for the simulation experiment
}
\keyword{internal}
