% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{add_slope}
\alias{add_slope}
\title{Add condition lines based on slope}
\usage{
add_slope(cond, cond_tbl)
}
\arguments{
\item{cond}{Character string that defines condition for the variable 'var'}

\item{cond_tbl}{Tibble with conditions defined based on soil and land use}
}
\value{
The conditions table with the slope row added.
}
\description{
The function translates the condition \code{cond} defined for the slope into a
tibble which can be added to the already existing conditions tibble \code{cond_tbl}.
}
\keyword{internal}
