% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{add_suffix_to_duplicate}
\alias{add_suffix_to_duplicate}
\title{Add suffix value to duplicated column names of SWAT+ output files..}
\usage{
add_suffix_to_duplicate(col_name)
}
\arguments{
\item{col_name}{Vector of column names}
}
\description{
Add suffix value to duplicated column names of SWAT+ output files..
}
\keyword{internal}
