% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{as_num}
\alias{as_num}
\title{Convert character string to numeric without displaying warnings}
\usage{
as_num(chr)
}
\arguments{
\item{chr}{Character string}
}
\description{
Convert character string to numeric without displaying warnings
}
\keyword{internal}
