% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{build_constraint_tbl}
\alias{build_constraint_tbl}
\title{Build the parameter constraint from the rules in the parameter names}
\usage{
build_constraint_tbl(cons_i, bool_op)
}
\arguments{
\item{cons_i}{Text string that defines the constraint i}

\item{bool_op}{Vector of strings that define the different possible boolean operations}
}
\description{
Build the parameter constraint from the rules in the parameter names
}
\keyword{internal}
