% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{build_expr}
\alias{build_expr}
\title{Build expresions used to filter parameter values in parameter modification
step}
\usage{
build_expr(var, val, op)
}
\arguments{
\item{var}{List of filter variables}

\item{val}{List of values for the respective filter variables}

\item{op}{List of logical operators for the respective filter variables}
}
\description{
Build expresions used to filter parameter values in parameter modification
step
}
\keyword{internal}

