% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{build_filter}
\alias{build_filter}
\title{Build filter expresions for one sequence of variable, operator and values}
\usage{
build_filter(var, val, op)
}
\arguments{
\item{var}{Character string providing the filter variable}

\item{val}{Character string providing the values for the respective filter
variable}

\item{op}{Character string providing the logical operator for the filter
operation}
}
\description{
Build filter expresions for one sequence of variable, operator and values
}
\keyword{internal}
