% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_utils.R
\name{build_scenario_run}
\alias{build_scenario_run}
\title{Prepare scenario thread paths}
\usage{
build_scenario_run(project_path, run_path, n_parallel)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{n_parallel}{Named vector whith number of cores to be used for parallel
threads of scenarios and parameters}
}
\value{
Character vector with the full thread paths
}
\description{
Prepare scenario thread paths
}
\keyword{internal}
