% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{check_dates}
\alias{check_dates}
\title{Check if input start and end dates are given in the weather inputs}
\usage{
check_dates(project_path, model_setup)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{model_setup}{list of model setup features}
}
\description{
Check if input start and end dates are given in the weather inputs
}
\keyword{internal}
