% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{check_for_end_date}
\alias{check_for_end_date}
\title{Check model output table if it contains the end date of the simulation}
\usage{
check_for_end_date(model_output, end_date)
}
\arguments{
\item{model_output}{model output list}

\item{end_date}{End date as character or numeric value}
}
\description{
Check model output table if it contains the end date of the simulation
}
\keyword{internal}
