% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{check_lock}
\alias{check_lock}
\title{Manages lock file for writing to the SQLite data bases to avoid simultanious
writing. This is advantageous for large files and a large number of parallel
threads.}
\usage{
check_lock(save_path)
}
\arguments{
\item{save_path}{Path to the folder that holds the saved data}
}
\description{
Manages lock file for writing to the SQLite data bases to avoid simultanious
writing. This is advantageous for large files and a large number of parallel
threads.
}
\keyword{internal}
