% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{check_output}
\alias{check_output}
\title{Check output if is a data.frame and convert in case to named list}
\usage{
check_output(output, swat_vers)
}
\arguments{
\item{output}{Defined output}
}
\description{
Check output if is a data.frame and convert in case to named list
}
\keyword{internal}
