% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_parameter}
\alias{check_parameter}
\title{Checks if all provided parameters exist in 'Absolute_SWAT_Values'}
\usage{
check_parameter(parameter, abs_swat_val)
}
\arguments{
\item{parameter}{Named parameter vector or parameter table}

\item{abs_swat_val}{(optional) path to custom 'Absolute_SWAT_Value' file}
}
\description{
Checks if all provided parameters exist in 'Absolute_SWAT_Values'
}
\keyword{internal}
