% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{check_revision}
\alias{check_revision}
\title{Check the Revision of used SWAT+ executable}
\usage{
check_revision(project_path, run_path, os, swat_exe)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{run_path}{Path where the '.model_run' folder is built. If NULL the
executable model is built in the 'project_path'}

\item{os}{Character string that gives the operationg system}

\item{swat_exe}{Character string that gives the name of the SWAT executable.}
}
\description{
Check the Revision of used SWAT+ executable
}
\keyword{internal}
