% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_run_index}
\alias{check_run_index}
\title{Checks the provided run_index vector in accordance with provided parameters}
\usage{
check_run_index(run_index, parameter)
}
\arguments{
\item{run_index}{Numeric vector to subset the provided parameter sets}

\item{parameter}{Named parameter vector or parameter table}
}
\description{
Checks the provided run_index vector in accordance with provided parameters
}
\keyword{internal}
