% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{check_saved_data}
\alias{check_saved_data}
\title{Do general checkups for a SQLite database that already holds saved data and
compare with the current run_swat inputs}
\usage{
check_saved_data(save_path, parameter, output, run_index, model_setup)
}
\arguments{
\item{save_path}{Path to the folder that holds the saved data}

\item{parameter}{List that provides Parameter set for the simualtion and the
parameter definition table}
}
\description{
Do general checkups for a SQLite database that already holds saved data and
compare with the current run_swat inputs
}
\keyword{internal}
