% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{check_swatplus_parameter}
\alias{check_swatplus_parameter}
\title{Check if the names of the defined parameters are available in 'cal_parms.cal'.}
\usage{
check_swatplus_parameter(project_path, parameter)
}
\arguments{
\item{project_path}{Path to the SWAT+ project}

\item{parameter}{Model parameter data set}
}
\description{
Check if the names of the defined parameters are available in 'cal_parms.cal'.
}
\keyword{internal}
