% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{cio_to_numeric}
\alias{cio_to_numeric}
\title{Helper function to convert file.cio entries to numerics}
\usage{
cio_to_numeric(cio_entry)
}
\arguments{
\item{cio_entry}{Line from file.cio}
}
\description{
Helper function to convert file.cio entries to numerics
}
\keyword{internal}
