% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{collect_cols}
\alias{collect_cols}
\title{Collect selected columns from table in SQLite database}
\usage{
collect_cols(col_name, tbl_name, db)
}
\arguments{
\item{col_name}{Character vector that defines the column nanems}

\item{tbl_name}{String that defines the name of the table from which
to collect the columns}

\item{db}{SQLite database connection}
}
\description{
Collect selected columns from table in SQLite database
}
\keyword{internal}
