% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{collect_sim_i}
\alias{collect_sim_i}
\title{Helper function to collect a table from a sqlite date base according to the
information given in one line of the table_overview}
\usage{
collect_sim_i(sim_i, save_list)
}
\arguments{
\item{sim_i}{One line of the overview_table providing with the columns
tbl_name (the name of a table in one of the sqlite db's), var (name of the
variable), run_label, run_num (the numeric representation of a run), and
the con_number (providing the number of the data base connection that
stores the results for this variable and run)}
}
\description{
Helper function to collect a table from a sqlite date base according to the
information given in one line of the table_overview
}
\keyword{internal}
