% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{collect_sim_run}
\alias{collect_sim_run}
\title{Wrapper for sim_i that collects all variables for a simulation run}
\usage{
collect_sim_run(sim_run)
}
\arguments{
\item{sim_i}{same as sim_i above but multiple lines where each line is a
variable for the run_i}
}
\description{
Wrapper for sim_i that collects all variables for a simulation run
}
\keyword{internal}
