% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{compare_parameter}
\alias{compare_parameter}
\title{Compare if parameters saved in the database and parameters of simulation
match}
\usage{
compare_parameter(par_val, par_def, par_val_db, par_def_db)
}
\arguments{
\item{par_val}{Parameter values table in simulation}

\item{par_def}{Parameter definition table in simulation}

\item{par_val_db}{Parameter values table in data base}

\item{par_def_db}{Parameter definition table in data base}
}
\description{
Compare if parameters saved in the database and parameters of simulation
match
}
\keyword{internal}
