% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{compare_tables}
\alias{compare_tables}
\title{Update a time stamp in the run_log table}
\usage{
compare_tables(tbl_sim, tbl_df, txt, convert = FALSE)
}
\arguments{
\item{tbl_sim}{Table defined for the current simulation}

\item{tbl_df}{Corresponding table saved in the data bases}

\item{txt}{Text string to be added in the error message.}

\item{save_path}{Path of the sql data base for incrementally saving simulations}

\item{run_index}{IDs of parameter sets for which simulations should be run}
}
\description{
Update a time stamp in the run_log table
}
\keyword{'save_file'}
\keyword{Compare}
\keyword{SWAT}
\keyword{a}
\keyword{current}
\keyword{in}
\keyword{internal}
\keyword{one}
\keyword{saved}
\keyword{simulation}
\keyword{table}
\keyword{the}
\keyword{with}
