% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{convert_date}
\alias{convert_date}
\title{Convert dates from date format to year/month/day/hour/min/sec columns or vice
versa}
\usage{
convert_date(date_tbl)
}
\arguments{
\item{date_tbl}{Table holding either one date column or the splitted date
columns}
}
\description{
Convert dates from date format to year/month/day/hour/min/sec columns or vice
versa
}
\keyword{internal}

