% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{define_output}
\alias{define_output}
\title{Define output variables}
\usage{
define_output(file, variable = NULL, unit = NULL, expression = NULL)
}
\arguments{
\item{variable}{The name of the variable as given in the header of the
respective output output_file (without unsits!)}

\item{...}{Variables that will be provided as output after the execution
of \code{run_SWAT()}. The variables are defined with the helper
function \code{variable()}. See the examples for possible application.}

\item{output_file}{The SWAT output output_file where the output variable is located (so
far only .sub and .rch are supported!)}

\item{spatial_unit}{The 'RCH' or 'SUB unit for which the output should be
written.}

\item{expr}{Caution! Advanced setting if the output variable should be
modified or aggregated before writing the output. To do so a piped
workflow must be provided (see examples).}

\item{add_date}{Logical parameter whether a date should be provided for
the output or not. E.g. can be set FALSE if temporal aggregation was
applied to the output variable and the date would have no meaning for
the variable.}
}
\value{
Returns a tibble of the output variables to provide to the
  parameter \code{output} of the function \code{run_SWAT()}

The function should only be used as a helper function for
  \code{define_output()}
}
\description{
Define which output variables (from the parallel SWAT execution of the
function \code{\link{run_swat}}) should be saved and returned after finishing
the model runs. Note that, the variable \code{output} in
\code{\link{run_swat}} must be defined via this function.
}
\examples{
# Writing the discharge for the RCH unsits 1 to 5
# and the nitrogen load and ET for the RCH or SUB unit 5:
  out_tbl <- define_output(flow = out_var(output_file = "rch",
                                           variable = "FLOW_OUT",
                                           spatial_unit = 1:5),
                           no3  = out_var(output_file = "rch",
                                           variable = "NO3_OUT",
                                           spatial_unit = 5),
                           ptot = out_var(output_file = "rch",
                                           variable = "PTOT",
                                           spatial_unit = 5),
                           et_a = out_var(output_file = "sub",
                                           variable = "ET",
                                           spatial_unit = 5))


Helper function for the \\code{\\link{run_swat}}) output variable definition with
\\code{\\link{define_output}}

Each variable provided with the function \\code{\\link{define_output}}should be
defined using the helper function \\code{\\link{variable}}.

}

