% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{display_date}
\alias{display_date}
\title{Convert the information on the dates for the simulated variables into strings
that are printed}
\usage{
display_date(date_data)
}
\arguments{
\item{date_data}{Tables holding the dates loaded from the sqlite data bases}
}
\description{
Convert the information on the dates for the simulated variables into strings
that are printed
}
\keyword{internal}

