% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{display_progress_pct}
\alias{display_progress_pct}
\title{Display the progress if iterative processes as percentage value}
\usage{
display_progress_pct(n, nmax, t0)
}
\arguments{
\item{n}{Iteration step}

\item{nmax}{Number of iterations}

\item{t0}{initial time step}
}
\description{
Display the progress if iterative processes as percentage value
}
\keyword{internal}
