% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{display_runs}
\alias{display_runs}
\title{Convert the information on available runs for the simulated variables into
strings that are printed}
\usage{
display_runs(runs)
}
\arguments{
\item{tbl}{overview table that provides meta data for all simulation runs for
all variables saved in the data bases}
}
\description{
Convert the information on available runs for the simulated variables into
strings that are printed
}
\keyword{internal}
