% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat2012_subfunctions.R
\name{evaluate_expression}
\alias{evaluate_expression}
\title{Evaluate the expression defined for a variable in 'output'}
\usage{
evaluate_expression(out_table, expression)
}
\arguments{
\item{expression}{Expression to be applied to extract variable from
output_table}

\item{output_table}{Output defined to read from the SWAT model results}
}
\description{
Evaluate the expression defined for a variable in 'output'
}
\keyword{internal}

