% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{extract_non_var_cols}
\alias{extract_non_var_cols}
\title{Extract date, plant_name, p columns from the respective simulation outputs}
\usage{
extract_non_var_cols(tbl)
}
\arguments{
\item{tbl}{Table of simulation results for one SWAT output file}
}
\description{
Extract date, plant_name, p columns from the respective simulation outputs
}
\keyword{internal}
