% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_outputs.R
\name{extract_out_i}
\alias{extract_out_i}
\title{Extract the variable i from the simulation outputs}
\usage{
extract_out_i(out_i, out_i_name, mod_out)
}
\arguments{
\item{out_i}{The ith output defined to read from the SWAT model results}

\item{out_i_name}{The name of the ith output variable}

\item{mod_out}{Output files read from the respective thread}
}
\description{
Extract the variable i from the simulation outputs
}
\keyword{internal}
