% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_outputs.R
\name{extract_output}
\alias{extract_output}
\title{Extract the variables from the model outputs as defined in 'output'}
\usage{
extract_output(output, model_output)
}
\arguments{
\item{output}{Output defined to read from the SWAT model results}

\item{model_output}{Output files read from the respective thread}
}
\description{
Extract the variables from the model outputs as defined in 'output'
}
\keyword{internal}
