% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{extract_swat2012_output_i}
\alias{extract_swat2012_output_i}
\title{Extract and rearrange output variables of SWAT2012 outputs}
\usage{
extract_swat2012_output_i(out_tbl_i, out_def_i)
}
\arguments{
\item{out_tbl_i}{ith output table read from SWAT2012 simulations}

\item{out_def_i}{ith output definition table}
}
\description{
Extract and rearrange output variables of SWAT2012 outputs
}
\keyword{internal}
