% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{filter_mon}
\alias{filter_mon}
\title{Filter function to extract only relevant month entries in SWAT2012 outputs}
\usage{
filter_mon(mon)
}
\arguments{
\item{mon}{Month column from output table as vector}
}
\description{
Filter function to extract only relevant month entries in SWAT2012 outputs
}
\keyword{internal}
