% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{filter_not_empty}
\alias{filter_not_empty}
\title{Filter the elements of a list with tibbles where the tibbles are not empty}
\usage{
filter_not_empty(dat_list)
}
\arguments{
\item{dat_list}{List of tibbles (data.frames)}
}
\description{
Filter the elements of a list with tibbles where the tibbles are not empty
}
\keyword{internal}
