% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_modify.R
\name{filter_spec_expr}
\alias{filter_spec_expr}
\title{Filter a parameter table for the specific expression if available}
\usage{
filter_spec_expr(par_table, expr)
}
\arguments{
\item{par_table}{Parameter table}

\item{expr}{Expression to evaluate}
}
\description{
Filter a parameter table for the specific expression if available
}
\keyword{internal}

