% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{find_duplicate}
\alias{find_duplicate}
\title{Find simulation runs that are duplicated in the provided sqlite data bases}
\usage{
find_duplicate(tbl)
}
\arguments{
\item{tbl}{overview table that provides meta data for all simulation runs for
all variables saved in the data bases}
}
\description{
Find simulation runs that are duplicated in the provided sqlite data bases
}
\keyword{internal}

