% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{find_exe_file}
\alias{find_exe_file}
\title{Find executable file in the project folder}
\usage{
find_exe_file(project_path, os)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder)}

\item{os}{Operating system of computer, one of 'win', 'unix'.}
}
\value{
the name of the executable file
}
\description{
Find executable file in the project folder
}
\keyword{internal}
