% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{find_first_line}
\alias{find_first_line}
\title{Helper function to find the position of the fist line of the table in a file}
\usage{
find_first_line(out_file, thread_path)
}
\arguments{
\item{out_file}{The output files to be read}

\item{thread_path}{Path to respective thread where SWAT was executed}
}
\description{
Helper function to find the position of the fist line of the table in a file
}
\keyword{internal}
