% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R,
%   R/swatplus_handle_output.R
\name{find_first_space}
\alias{find_first_space}
\alias{find_first_space}
\title{Helper function to find the fist position of white spaces in a text string}
\usage{
find_first_space(string)

find_first_space(string)
}
\arguments{
\item{string}{text string that contains words seperated by white spaces}

\item{string}{text string that contains words seperated by white spaces}
}
\description{
Helper function to find the fist position of white spaces in a text string

Helper function to find the fist position of white spaces in a text string
}
\keyword{internal}
