% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_screen_parameter.R
\name{find_sensitivity_threshold}
\alias{find_sensitivity_threshold}
\title{Find threshold between informative and non-informative parameters based on
their sensitivities}
\usage{
find_sensitivity_threshold(sens_value)
}
\arguments{
\item{sens_value}{Numeric vector of parameter sensitivities}
}
\description{
Find threshold between informative and non-informative parameters based on
their sensitivities
}
\keyword{internal}
