% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{finish_progress}
\alias{finish_progress}
\title{Print message for completed process}
\usage{
finish_progress(nmax, t0, word)
}
\arguments{
\item{nmax}{Number of iterations}

\item{t0}{initial time step}
}
\description{
Print message for completed process
}
\keyword{internal}
