% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_write.R
\name{format_mgt_line}
\alias{format_mgt_line}
\title{Format a line in the mgt table for writing in correct format to file}
\usage{
format_mgt_line(mgt_line)
}
\arguments{
\item{mgt_line}{A line in the mgt table to be formatted as character strings}
}
\description{
Format a line in the mgt table for writing in correct format to file
}
\keyword{internal}
