% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{format_swat2012_parameter}
\alias{format_swat2012_parameter}
\title{Translate the parameter inputs into a parameter input table and a separarate
table providing the file constraints and the filter expressions for the
respective parameter}
\usage{
format_swat2012_parameter(parameter, swat_vers)
}
\arguments{
\item{parameter}{Model parameters as named vector or tibble}
}
\description{
Translate the parameter inputs into a parameter input table and a separarate
table providing the file constraints and the filter expressions for the
respective parameter
}
\keyword{internal}
