% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{format_swatplus_parameter}
\alias{format_swatplus_parameter}
\title{Translate the parameter inputs into a parameter input table and a separarate
table providing the file constraints and the filter expressions for the
respective parameter}
\usage{
format_swatplus_parameter(parameter)
}
\arguments{
\item{parameter}{Model parameters as named vector or tibble}
}
\description{
Translate the parameter inputs into a parameter input table and a separarate
table providing the file constraints and the filter expressions for the
respective parameter
}
\keyword{internal}
