% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_outputs.R
\name{get_date_vector}
\alias{get_date_vector}
\title{Create date vector from the date info in the model setup}
\usage{
get_date_vector(model_setup)
}
\arguments{
\item{model_setup}{Model setup defined by run_swat input parameters and/of
the SWAT model input files}
}
\description{
Create date vector from the date info in the model setup
}
\keyword{internal}
