% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{get_date_vector_2012}
\alias{get_date_vector_2012}
\title{Create date vector from the date info in the model setup of a SWAT2012 project}
\usage{
get_date_vector_2012(model_setup)
}
\arguments{
\item{model_setup}{Model setup defined by run_swat input parameters and/of
the SWAT model input files}
}
\description{
Create date vector from the date info in the model setup of a SWAT2012 project
}
\keyword{internal}
