% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_outputs.R
\name{get_date_vector_plus}
\alias{get_date_vector_plus}
\title{Create date vector from the date columns in the last SWAT+ simulation}
\usage{
get_date_vector_plus(output, thread_path, model_setup, revision)
}
\arguments{
\item{output}{Output definition that results from \code{define_output()}}

\item{thread_path}{Chracter string that defines the path to thread_1
the SWAT model input files}

\item{model_setup}{Model setup defined by run_swat input parameters and/of}

\item{revision}{Numeric value that defines the SWAT+ revision}
}
\description{
Create date vector from the date columns in the last SWAT+ simulation
}
\keyword{internal}
