% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_digit}
\alias{get_digit}
\title{Get the number of digits that the run_XXX should have based on the number
of rows in the parameter values table}
\usage{
get_digit(tbl)
}
\arguments{
\item{tbl}{Parameter values table}
}
\description{
Get the number of digits that the run_XXX should have based on the number
of rows in the parameter values table
}
\keyword{internal}
