% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{get_hru_meta}
\alias{get_hru_meta}
\title{Extract the meta data from the header of a '.hru' file}
\usage{
get_hru_meta(hru_file_i)
}
\arguments{
\item{hru_file_i}{The i'th hru file (read as character vector)}
}
\description{
Extract the meta data from the header of a '.hru' file
}
\keyword{internal}
