% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{get_n_parallel}
\alias{get_n_parallel}
\title{Get numbers of parallel threads for scenarios and parameters which minimize
the number of parallel iterations.}
\usage{
get_n_parallel(n_thread, n_scenario, n_parameter)
}
\arguments{
\item{n_thread}{Number of parallel threads defined by the user.}

\item{n_scenario}{Number of scenarios which should be run.}

\item{n_parameter}{Number of parameter sets which should be run for every
scenario.}
}
\value{
A vector of length 2 with the number of parallel threads for
scenarios and parameters.
}
\description{
Get numbers of parallel threads for scenarios and parameters which minimize
the number of parallel iterations.
}
\keyword{internal}
