% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_utils.R
\name{get_n_parameter}
\alias{get_n_parameter}
\title{Get the number of parameter sets which shold be run for each scenario.}
\usage{
get_n_parameter(parameter = NULL, run_index = NULL, ...)
}
\arguments{
\item{parameter}{Parameter input table that is passed as an optional
\verb{run_swat*()} input argument.}

\item{run_index}{Indices of parameter combintations which should be run
from the parameter set in \code{parameter}. This is an optional input argument
passed on to \verb{run_swat*()}.}

\item{...}{Addtional input arguments which can be passed on. Not considered
in this function.}
}
\value{
the number of parameter sets as integer value
}
\description{
Get the number of parameter sets which shold be run for each scenario.
}
\keyword{internal}
