% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{get_par_name}
\alias{get_par_name}
\title{Extract the parameter name from one line in a list parameter file}
\usage{
get_par_name(chr, par_pos)
}
\arguments{
\item{chr}{Character string}

\item{par_pos}{Logical vector that provides if a row holds a parameter or not}
}
\description{
Extract the parameter name from one line in a list parameter file
}
\keyword{internal}
