% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{get_sol_texture}
\alias{get_sol_texture}
\title{Acquire the object indices of from the respective object file.}
\usage{
get_sol_texture(file)
}
\arguments{
\item{file}{Path to the object file}
}
\description{
Acquire the object indices of from the respective object file.
}
\keyword{internal}
