% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{get_table}
\alias{get_table}
\title{Extract the parameter values that are provided in tabular form in a file}
\usage{
get_table(file_i, table_pos, col_pos, col_names, fun)
}
\arguments{
\item{file_i}{The i'th parameter file for a file suffix}

\item{table_pos}{Index vector indicating which lines belong to the table}

\item{col_pos}{Index vector that separates the individual table columns}
}
\description{
Extract the parameter values that are provided in tabular form in a file
}
\keyword{internal}
