% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{get_tbl_column}
\alias{get_tbl_column}
\title{Acquire the object indices of from the respective object file.}
\usage{
get_tbl_column(file, col_i)
}
\arguments{
\item{file}{Path to the object file}
}
\description{
Acquire the object indices of from the respective object file.
}
\keyword{internal}
