% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_time_interval}
\alias{get_time_interval}
\title{Get the time interval between to time stamps}
\usage{
get_time_interval(start_time, end_time)
}
\arguments{
\item{start_time}{start time stamp}

\item{end_time}{end time stamp}
}
\description{
Get the time interval between to time stamps
}
\keyword{internal}
