% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{get_value}
\alias{get_value}
\title{Extract the parameter values from a list parameter file}
\usage{
get_value(file_i, par_pos)
}
\arguments{
\item{file_i}{The i'th parameter file for a file suffix}

\item{par_pos}{Logical vector that provides if a row holds a parameter or not}
}
\description{
Extract the parameter values from a list parameter file
}
\keyword{internal}
