% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{group_units}
\alias{group_units}
\title{Group sequences of units}
\usage{
group_units(val)
}
\arguments{
\item{val}{Numeric vector with the unit values}
}
\value{
A numeric vector with the required syntax (e.g. 1 to 5 and 8 is
c(1, -5, 8))
}
\description{
The unit condition in calibration.cal follows a specific syntax. This function
translates a vector of units into groups with that syntax for writing the OBJ
columns in calibration.cal
}
\keyword{internal}
