% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{group_variable_units}
\alias{group_variable_units}
\title{Group variable names which where saved for several spatial units}
\usage{
group_variable_units(variable_names)
}
\arguments{
\item{variable_names}{Vector with variable names}
}
\description{
Group variable names which where saved for several spatial units
}
\keyword{internal}
