% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{identify_sequence}
\alias{identify_sequence}
\title{Group sequences of units together for writing the OBJ columns in calibration.cal}
\usage{
identify_sequence(val)
}
\arguments{
\item{val}{Numeric vector with the unit values}
}
\description{
Group sequences of units together for writing the OBJ columns in calibration.cal
}
\keyword{internal}
