% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{init_calibration_cal}
\alias{init_calibration_cal}
\title{Initialize a line in the calibration.cal table for the parameter i}
\usage{
init_calibration_cal(par_def_i)
}
\arguments{
\item{par_def_i}{Definition of parameter i (i.e. the ith line of parameter$definition).}
}
\value{
A tibble with one line which initializes the parameter i.
}
\description{
Initialize a line in the calibration.cal table for the parameter i
}
\keyword{internal}
