% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{is_identical_lst}
\alias{is_identical_lst}
\title{Check if tables in a list are identical}
\usage{
is_identical_lst(tbl_list)
}
\arguments{
\item{tbl_list}{List of data.frames}
}
\description{
Check if tables in a list are identical
}
\keyword{internal}
