% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{is_identical_tbl}
\alias{is_identical_tbl}
\title{Check if two tables are identical}
\usage{
is_identical_tbl(x, y)
}
\arguments{
\item{x}{First table}

\item{y}{Second table}
}
\description{
Check if two tables are identical
}
\keyword{internal}
