% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{is_par}
\alias{is_par}
\title{Check which row in a character vector file holds a model parameter}
\usage{
is_par(chr)
}
\arguments{
\item{chr}{Character string}
}
\description{
Check which row in a character vector file holds a model parameter
}
\keyword{internal}
