% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_handling.R
\name{load_demo}
\alias{load_demo}
\title{Load SWAT demo data}
\usage{
load_demo(dataset, swat_version = NULL, path = NULL)
}
\arguments{
\item{path}{Character string that defines the path where copy the SWAT demo
project.}

\item{type}{Character string that defines the type of demo data set to be
loaded. \code{type = 'SWAT2012'} loads a SWAT2012 project folder.
\code{type = 'SWATplus'} loads a SWAT+ project folder. \code{type =
'obs_data'} returns the observation data for the SWAT demo porject.}
}
\description{
Load SWAT demo data
}
