% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{load_swat_run}
\alias{load_swat_run}
\title{Load saved SWAT simulations}
\usage{
load_swat_run(
  save_dir,
  variable = NULL,
  run = NULL,
  add_parameter = TRUE,
  add_date = TRUE,
  add_run_info = TRUE
)
}
\arguments{
\item{save_dir}{Character string or vector of character strings that provide
the path/s to the save folder/s.}

\item{variable}{Output variables that were saved in the SWAT run and that
should be loaded into R}

\item{run}{Numeric vector giving the indexes of the simulations that should
be loaded}

\item{add_parameter}{Logical. If \code{add_parameter = TRUE} the parameter
set for the SWAT runs is added to the loaded data}

\item{add_date}{Logical. If \code{add_date = TRUE} a date column is added to
the simulation results of each variable}

\item{add_run_info}{Logical. If \code{add_run_info = TRUE} the meta information
for the simulation runs is added to the outputs}
}
\description{
Load simuation results that were saved in sqlite data bases in one or
multiple save folders when running SWAT.
}
