% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_screen_parameter.R
\name{logi_curv}
\alias{logi_curv}
\title{Curvature of the logistic regression function}
\usage{
logi_curv(x, par)
}
\arguments{
\item{x}{Numeric vector of x values}

\item{par}{Numeric parameter set, providing (1) offset, (2) Curve's max
point, (3) steepness, and (4) x value of sigmoid mid-point}
}
\description{
Curvature of the logistic regression function
}
\keyword{internal}
