% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_modify.R
\name{modify_mgt_par}
\alias{modify_mgt_par}
\title{Modify parameters that are provided in the management table format}
\usage{
modify_mgt_par(parameter, model_parameter, file_meta, i_par, run_index, i_run)
}
\arguments{
\item{parameter}{List providing the parameter table and the parameter
constraints}

\item{model_parameter}{The swat_parameter set that will be modified according
to the parpameter i_par}

\item{file_meta}{Table with the meta informations on the swat model parameter
files}

\item{i_par}{Index that gives the number of the curent parameter that is
modified in this step}

\item{run_index}{Vector of the indices of runs that are performed}

\item{i_run}{Index that gives the number of the current run simulated in the
respective thread}
}
\description{
Modify parameters that are provided in the management table format
}
\keyword{internal}
