% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{mutate_output_i}
\alias{mutate_output_i}
\title{Transform extracted outputs into a wide table and add suffix to variable names.}
\usage{
mutate_output_i(out_tbl_i, output_i)
}
\arguments{
\item{out_tbl_i}{i_th output table read from the SWAT+ outputs}

\item{output_i}{i_th part from the output table which defines what to
read from the SWAT model results}
}
\description{
Transform extracted outputs into a wide table and add suffix to variable names.
}
\keyword{internal}
