% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{paste_runs}
\alias{paste_runs}
\title{Paste run indexes if start and end of sequence differ. Otherwise only use
start value}
\usage{
paste_runs(strt, end, sep)
}
\arguments{
\item{strt}{Numeric start value of sequence}

\item{end}{Numeric end value of sequence}
}
\description{
Paste run indexes if start and end of sequence differ. Otherwise only use
start value
}
\keyword{internal}
