% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_dotty}
\alias{plot_dotty}
\title{Dotty for an objective criterion over all parameters}
\usage{
plot_dotty(parameter, objective)
}
\arguments{
\item{parameter}{data.frame that provides all paramete sets used for simulation}

\item{objective}{Vector or data.frame providing all calculated objective criteria}
}
\value{
Interactive visualization of observed (blue) and simulation
results (red). The red ribbon gives the upper and lower boundary for each
time step from the selected model runs.
}
\description{
Dotty for an objective criterion over all parameters
}

