% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Interactive timeseries plot of the simulation results filtered by a criterion}
\usage{
plot_timeseries(sim, obs, obj, crit, thres)
}
\arguments{
\item{sim}{Simulation results for the SWAT model runs}

\item{obs}{data.frame holding the observed data. First column must be of
type date. Second column must hold numeric observations for respective dates.}

\item{obj}{Vector giving the values of the objective calculated for the
individual model runs. Must be of same length as the number of simulations.}

\item{crit}{Criterion applied. Either ">" for model runs with objective
values greater than thres or "<" for smaller values to be selected.}

\item{thres}{Threshold value that is applied for selection.}
}
\value{
Interactive visualization of observed (blue) and simulation
results (red). The red ribbon gives the upper and lower boundary for each
time step from the selected model runs.
}
\description{
Interactive timeseries plot of the simulation results filtered by a criterion
}

