% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{prepare_error_report}
\alias{prepare_error_report}
\title{Prepare error report in case of failed simulations}
\usage{
prepare_error_report(sim_result)
}
\arguments{
\item{sim_result}{List of simulation results from the SWAT model runs}
}
\description{
Prepare error report in case of failed simulations
}
\keyword{internal}
