% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{prepare_run_info}
\alias{prepare_run_info}
\title{Prepare run info for the simulation experiment}
\usage{
prepare_run_info(
  sim_result,
  model_setup,
  output,
  run_index,
  project_path,
  t0,
  t1
)
}
\arguments{
\item{model_setup}{List of model configurations}

\item{output}{Table of defined output variables}

\item{run_index}{Vector of run IDs}

\item{project_path}{Path to the SWAT project folder}

\item{t0}{Start date time stamp}

\item{t1}{End date time stamp}

\item{sim_results}{List of simulation results from the SWAT model runs}
}
\description{
Prepare run info for the simulation experiment
}
\keyword{internal}
