% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_basin_yld}
\alias{read_basin_yld}
\title{Read basin yield output tables.}
\usage{
read_basin_yld(output_i, thread_path)
}
\arguments{
\item{output_i}{i_th part from the output table which defines what to}

\item{thread_path}{String path to the thread where to read the output file
read from the SWAT model results}
}
\description{
Read basin yield output tables.
}
\keyword{internal}
