% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat2012_subfunctions.R
\name{read_date}
\alias{read_date}
\title{Read file.cio parameters and convert the required ones to a date sequence}
\usage{
read_date(file_cio)
}
\arguments{
\item{file_cio}{Modified file.cio}
}
\description{
Read file.cio parameters and convert the required ones to a date sequence
}
\keyword{internal}

