% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_fdcout}
\alias{read_fdcout}
\title{Reading the FDC output table.}
\usage{
read_fdcout(output_i, thread_path)
}
\arguments{
\item{output_i}{i_th part from the output table which defines what to
read from the SWAT model results}

\item{thread_path}{String path to the thread where to read the output file}
}
\description{
Reading the FDC output table.
}
\keyword{internal}
