% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_file_meta}
\alias{read_file_meta}
\title{Read the meta information for the parameter files}
\usage{
read_file_meta(project_path, par_constrain)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Read the meta information for the parameter files
}
\keyword{internal}
