% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_hru}
\alias{read_hru}
\title{Read all '.hru' files in the project_path and extract the meta data from
these files and return them as a meta data tibble}
\usage{
read_hru(project_path)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Read all '.hru' files in the project_path and extract the meta data from
these files and return them as a meta data tibble
}
\keyword{internal}
