% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_mgtout}
\alias{read_mgtout}
\title{Read SWAT+ management output file and return the read output in a tibble}
\usage{
read_mgtout(output_i, thread_path)
}
\arguments{
\item{run_path}{Path to the folder where simulations are performed}
}
\description{
Read SWAT+ management output file and return the read output in a tibble
}
\keyword{internal}
