% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat2012_subfunctions.R
\name{read_output}
\alias{read_output}
\title{Read SWAT output files}
\usage{
read_output(ouput, thread_path)
}
\arguments{
\item{thread_path}{Path to respective thread where SWAT was executed}

\item{output}{Output defined to read from the SWAT model results}
}
\description{
Read SWAT output files
}
\keyword{internal}

