% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_output_i}
\alias{read_output_i}
\title{Reading the i_th SWAT+ output file, filter required units and select variables.}
\usage{
read_output_i(output_i, col_names_i, thread_path, date_cols)
}
\arguments{
\item{output_i}{i_th part from the output table which defines what to
read from the SWAT model results}

\item{col_names_i}{Prepared and fixed column names of output file i}

\item{thread_path}{String path to the thread where to read the output file}

\item{date_cols}{If data should be read, vector of names of the date columns.}
}
\description{
Reading the i_th SWAT+ output file, filter required units and select variables.
}
\keyword{internal}
