% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_par_files}
\alias{read_par_files}
\title{Read the original parameter values from the parameter files in project_path}
\usage{
read_par_files(project_path)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Read the original parameter values from the parameter files in project_path
}
\keyword{internal}
