% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_par_list}
\alias{read_par_list}
\title{Read parameters that are arranged in a simple list (1 parameter per row)}
\usage{
read_par_list(file_meta, file_suffix, project_path, n_row = NULL)
}
\arguments{
\item{file_meta}{Table that provides the file meta data}

\item{file_suffix}{Suffix of the parameter files from which parameters are
read}

\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Read parameters that are arranged in a simple list (1 parameter per row)
}
\keyword{internal}
