% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{read_sol}
\alias{read_sol}
\title{Read parameters from the '.sol' file (as it has an individual file structure)}
\usage{
read_sol(file_meta, project_path)
}
\arguments{
\item{file_meta}{Table that provides the file meta data}

\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Read parameters from the '.sol' file (as it has an individual file structure)
}
\keyword{internal}
