% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{read_swat2012_date}
\alias{read_swat2012_date}
\title{Read file.cio parameters and convert the required ones to a date sequence}
\usage{
read_swat2012_date(file_cio)
}
\arguments{
\item{file_cio}{Modified file.cio}
}
\description{
Read file.cio parameters and convert the required ones to a date sequence
}
\keyword{internal}
