% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{read_swat2012_output}
\alias{read_swat2012_output}
\title{Read SWAT output files}
\usage{
read_swat2012_output(output, thread_path)
}
\arguments{
\item{output}{Output defined to read from the SWAT model results}

\item{thread_path}{Path to respective thread where SWAT was executed}
}
\description{
Read SWAT output files
}
\keyword{internal}
