% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_swatplus_date}
\alias{read_swatplus_date}
\title{Read date from first output file and return it for adding to the simulation
outputs}
\usage{
read_swatplus_date(output, run_path)
}
\arguments{
\item{output}{List of output variables defined with \code{define_output}}

\item{run_path}{Path to the simulation folder structure}
}
\description{
Read date from first output file and return it for adding to the simulation
outputs
}
\keyword{internal}
