% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_swatplus_output}
\alias{read_swatplus_output}
\title{Read SWAT+ output files}
\usage{
read_swatplus_output(output, thread_path, add_date, revision)
}
\arguments{
\item{output}{Output defined to read from the SWAT model results}

\item{thread_path}{Path to respective thread where SWAT was executed}

\item{add_date}{Logical wheter to add date column}

\item{revision}{Numeric. The revision number of the SWAT+ executable}
}
\description{
Read SWAT+ output files
}
\keyword{internal}
