% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{remove_dummy}
\alias{remove_dummy}
\title{Remove the generated dummy column that was required for the col bind}
\usage{
remove_dummy(tbl)
}
\arguments{
\item{tbl}{Tibble}
}
\description{
Remove the generated dummy column that was required for the col bind
}
\keyword{internal}
