% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{remove_non_var_cols}
\alias{remove_non_var_cols}
\title{Extract variable columns from the respective simulation outputs}
\usage{
remove_non_var_cols(tbl)
}
\arguments{
\item{tbl}{Table of simulation results for one SWAT output file}
}
\description{
Extract variable columns from the respective simulation outputs
}
\keyword{internal}
