% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{remove_units}
\alias{remove_units}
\title{Remove the units from the column names}
\usage{
remove_units(tbl)
}
\arguments{
\item{tbl}{table for which the units are removed from the col_names}
}
\description{
Remove the units from the column names
}
\keyword{internal}

