% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{remove_units_2012}
\alias{remove_units_2012}
\title{Remove the units from variable names in output files of SWAT2012 simulations}
\usage{
remove_units_2012(col_nm)
}
\arguments{
\item{col_nm}{Character vector with column names}
}
\description{
Remove the units from variable names in output files of SWAT2012 simulations
}
\keyword{internal}
