% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{remove_units_plus}
\alias{remove_units_plus}
\title{Remove the units from variable names in output files of SWAT+ Revisions before 56}
\usage{
remove_units_plus(col_nm)
}
\arguments{
\item{col_nm}{Character vector with column names}
}
\description{
Remove the units from variable names in output files of SWAT+ Revisions before 56
}
\keyword{internal}
