% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{replace_colname_na}
\alias{replace_colname_na}
\title{Fix issues with shifted col_names in SWAT+ rev59.3}
\usage{
replace_colname_na(col_nm, unit_nm)
}
\arguments{
\item{col_nm}{Character vector with column names}

\item{unit_nm}{Character vector with unit line from table}
}
\description{
Fix issues with shifted col_names in SWAT+ rev59.3
}
\keyword{internal}
