% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{run_os}
\alias{run_os}
\title{Add './' to run the exe on unix systems}
\usage{
run_os(exe, os)
}
\arguments{
\item{exe}{Text string that defines the name of the executable file}

\item{os}{Text string that defines the operating system}
}
\description{
Add './' to run the exe on unix systems
}
\keyword{internal}
