% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{save_error_log}
\alias{save_error_log}
\title{Save error report of model run i in error_log sql data base}
\usage{
save_error_log(save_path, model_output, parameter, run_index, i_run)
}
\arguments{
\item{save_path}{Path of the sql data base}

\item{model_output}{output of the i_run'th simulation as a tibble}

\item{parameter}{Vector or tibble with parameter sets}

\item{run_index}{Vector of the indices of runs that are performed}

\item{i_run}{The i'th run of the SWAT simulation}
}
\description{
Save error report of model run i in error_log sql data base
}
\keyword{internal}
