% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{scan_save_files}
\alias{scan_save_files}
\title{Scan the save folders of a SWAT run and return all meta data of this
simulation}
\usage{
scan_save_files(save_dir)
}
\arguments{
\item{save_dir}{Character string or vector of character strings that provide
the path/s to the save folder/s.}
}
\description{
Scan the save folders of a SWAT run and return all meta data of this
simulation
}
\keyword{internal}
