% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{scan_swat_run}
\alias{scan_swat_run}
\title{Retrieve information on saved SWAT runs}
\usage{
scan_swat_run(save_dir)
}
\arguments{
\item{save_dir}{Character string or vector}
}
\description{
Scan one or multiple save folders that belong to the same SWAT simulation and
get information on the simulation period, simulated variables and used
parameter sets.
}
