% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_read.R
\name{set_colnames}
\alias{set_colnames}
\title{Set the column names of a matrix}
\usage{
set_colnames(mtx, col_names)
}
\arguments{
\item{mtx}{The matrix}

\item{col_names}{Character vector with column names}
}
\description{
Set the column names of a matrix
}
\keyword{internal}
