% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{set_save_path}
\alias{set_save_path}
\title{Set the save path to the sqlite data base file}
\usage{
set_save_path(project_path, save_path, save_dir)
}
\arguments{
\item{project_path}{Character string. Path of SWAT project}

\item{save_path}{(optional) character string. save path if different to
project path}

\item{save_dir}{character string. Name of the sqlite data base directory}
}
\description{
Set the save path to the sqlite data base file
}
\keyword{internal}
