% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_handle_output.R
\name{split_by_units}
\alias{split_by_units}
\title{Split header line at the positions of units and return tidy header}
\usage{
split_by_units(header)
}
\arguments{
\item{header}{Character string header line}
}
\description{
Split header line at the positions of units and return tidy header
}
\keyword{internal}
