% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{split_out_tbl}
\alias{split_out_tbl}
\title{Split tables with more than ncol_max columns into list of tables}
\usage{
split_out_tbl(tbl, ncol_max)
}
\arguments{
\item{tbl}{Table which should be split}

\item{ncol_max}{Maximum number of columns before split is done}
}
\description{
Split tables with more than ncol_max columns into list of tables
}
\keyword{internal}
