% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_screen_parameter.R
\name{sse}
\alias{sse}
\title{Sum of squared errors}
\usage{
sse(x, y, par, fun)
}
\arguments{
\item{x}{Numeric vector of x values}

\item{y}{Numeric vector of 'observed' y values}

\item{par}{Numeric vector of parameter set}

\item{fun}{Model function to evaluate}
}
\description{
Sum of squared errors
}
\keyword{internal}
