% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{tidy_constraint}
\alias{tidy_constraint}
\title{Apply a set of operations to the rule strings to check and clean the defined operations}
\usage{
tidy_constraint(chr, bool_op)
}
\arguments{
\item{chr}{Text string that defines the constraint i}

\item{bool_op}{Vector of strings that define the different possible boolean operations}
}
\description{
Apply a set of operations to the rule strings to check and clean the defined operations
}
\keyword{internal}
