% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{tidy_simulations}
\alias{tidy_simulations}
\title{Tidy up and rearrange simulation results before returning them}
\usage{
tidy_simulations(sim_result)
}
\arguments{
\item{sim_result}{List of simulation results from the SWAT model runs}
}
\description{
Tidy up and rearrange simulation results before returning them
}
\keyword{internal}
