% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_constrain.R
\name{translate_parameter_constraints}
\alias{translate_parameter_constraints}
\title{Translate the parameter inputs and the set constraints into lookup table}
\usage{
translate_parameter_constraints(par, swat_vers)
}
\arguments{
\item{par}{Character string vector providing parameters and constraints}
}
\description{
Translate the parameter inputs and the set constraints into lookup table
}
\keyword{internal}
