% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{translate_rev55}
\alias{translate_rev55}
\title{Translate the output file settings defined according to print.prt to the
actual output file names or SWAT+ Revisions before 56}
\usage{
translate_rev55(output, output_interval)
}
\arguments{
\item{output}{List of output variables defined with \code{define_output}}

\item{output_interval}{the time interval for writing the simulated outputs}
}
\description{
Translate the output file settings defined according to print.prt to the
actual output file names or SWAT+ Revisions before 56
}
\keyword{internal}
