% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{truncate}
\alias{truncate}
\title{Truncate long character vectors}
\usage{
truncate(x, n, side = "left")
}
\arguments{
\item{x}{Character vector}

\item{n}{Threshold value when truncation should be done}

\item{side}{side of truncation with ... Either left or both}
}
\description{
Truncate long character vectors
}
\keyword{internal}
