% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{unit_output_to_wide}
\alias{unit_output_to_wide}
\title{Convert output tables with unit column to wide table with variable name and
unit}
\usage{
unit_output_to_wide(model_output_i)
}
\arguments{
\item{model_output_i}{Table with model output variables and unit column}
}
\description{
Convert output tables with unit column to wide table with variable name and
unit
}
\keyword{internal}
