% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_modify.R
\name{update_par}
\alias{update_par}
\title{Update parameter values according to the different options 'abs', 'rep', and
'rel'}
\usage{
update_par(par, par_up, change)
}
\arguments{
\item{par}{Vector with parameter values that are updated}

\item{par_up}{Value that is applied to the paramter values for updating}

\item{change}{Type of change, either: abs', 'rep', or 'rel'}
}
\description{
Update parameter values according to the different options 'abs', 'rep', and
'rel'
}
\keyword{internal}
