% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{update_plant_par}
\alias{update_plant_par}
\title{Modify plants.plt parameters}
\usage{
update_plant_par(thread_path, parameter, is_plant_par, run_index, i_run)
}
\arguments{
\item{thread_path}{Path to the parallel thread folder}

\item{parameter}{List providing the parameter table and the parameter
constraints}

\item{is_plant_par}{Logical vector that defines the plant parameters}

\item{run_index}{Vector of the indices of runs that are performed}

\item{i_run}{Index that gives the number of the current run simulated in the
respective thread}
}
\description{
Modify plants.plt parameters
}
\keyword{internal}
