% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{update_sim_log}
\alias{update_sim_log}
\title{Update a time stamp in the run_log table}
\usage{
update_sim_log(save_path, run_info)
}
\arguments{
\item{save_path}{Path of the sql data base for incrementally saving simulations}

\item{run_index}{IDs of parameter sets for which simulations should be run}
}
\description{
Update a time stamp in the run_log table
}
\keyword{internal}
