% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_parameter_write.R
\name{write_calibration}
\alias{write_calibration}
\title{Update the calibration file structure with the parameter set of the current
simulation run_i}
\usage{
write_calibration(thread_path, parameter, calibration, run_index, i_run)
}
\arguments{
\item{thread_path}{Path to the current parallel thread 'thread_i'}

\item{parameter}{Model parameters as named vector or tibble}

\item{calibration}{Template table structure of the calibration file}

\item{i_run}{Index of the i_th simulation run}
}
\description{
Update the calibration file structure with the parameter set of the current
simulation run_i
}
\keyword{internal}
