% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat2012_subfunctions.R
\name{write_model_in}
\alias{write_model_in}
\title{Write the Model.in file required for rewriting paramters with SWAT_Edit.exe'}
\usage{
write_model_in(parameter, thread_path, i_run)
}
\arguments{
\item{parameter}{Named parameter vector or parameter table}

\item{thread_path}{Path to respective thread where parameters are rewritten}

\item{i_run}{Index of current run to select the current parameter set}
}
\description{
Write the Model.in file required for rewriting paramters with SWAT_Edit.exe'
}
\keyword{internal}

