% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_write.R
\name{write_par_list}
\alias{write_par_list}
\title{Write modified SWAT parameter files that are arranged in a linear format}
\usage{
write_par_list(file_meta, file_suffix, thread_parameter, thread_path)
}
\arguments{
\item{file_meta}{Table that provides the parameter file meta data}

\item{file_suffix}{File ending of the respective parameter files to be
written}

\item{thread_parameter}{Modified model parameters to be written to respective
thread folder}

\item{thread_path}{Path to the thread folder where model is executed}
}
\description{
Write modified SWAT parameter files that are arranged in a linear format
}
\keyword{internal}
