% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_write.R
\name{write_parameter}
\alias{write_parameter}
\title{Write modified SWAT parameter files in respective thread folder}
\usage{
write_parameter(file_meta, thread_parameter, thread_path)
}
\arguments{
\item{file_meta}{Table that provides the parameter file meta data}

\item{thread_parameter}{Modified model parameters to be written to respective
thread folder}

\item{thread_path}{Path to the thread folder where model is executed}
}
\description{
Write modified SWAT parameter files in respective thread folder
}
\keyword{internal}
