% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{write_run_files}
\alias{write_run_files}
\title{Write the updated file.cio to all parallel folders}
\usage{
write_run_files(run_path, run_files)
}
\arguments{
\item{run_path}{Path to the .model_run folder}

\item{file_cio}{Updated file_cio to be written}
}
\description{
Write the updated file.cio to all parallel folders
}
\keyword{internal}
