% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_outputs.R
\name{add_run_info}
\alias{add_run_info}
\title{Prepare run info for the simulation experiment}
\usage{
add_run_info(run_info, sim_result, run_index)
}
\arguments{
\item{run_info}{Existing list of run information}

\item{sim_result}{List of simulation results from the SWAT model runs}

\item{run_index}{Vector of run IDs}
}
\description{
Prepare run info for the simulation experiment
}
\keyword{internal}
