% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{add_soil_luse}
\alias{add_soil_luse}
\title{Add condition lines based on soil and land use variables}
\usage{
add_soil_luse(cond, var, cond_all)
}
\arguments{
\item{cond}{Character string that defines condition for the variable 'var'}

\item{var}{Character string. Variable for which condition is applied}

\item{cond_all}{List of vectors that define all possible values for the
condition variables.}
}
\description{
Add condition lines based on soil and land use variables
}
\keyword{internal}
