% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{add_value_range}
\alias{add_value_range}
\title{Add the value range 'val' for the condition variable 'var' for a parameter to
calibration.cal}
\usage{
add_value_range(par_cal, val, var)
}
\arguments{
\item{par_cal}{The calibration.cal tibble for the parameter i}

\item{val}{Vector that defines the value range}

\item{var}{Character string. Name of the variable}
}
\description{
Add the value range 'val' for the condition variable 'var' for a parameter to
calibration.cal
}
\keyword{internal}
