% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{build_model_run}
\alias{build_model_run}
\title{Generate folder structure for parallel SWAT execution}
\usage{
build_model_run(project_path, run_path, n_thread, os, swat_vers, quiet)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{run_path}{Path where the '.model_run' folder is built. If NULL the
executable model is built in the 'project_path'}

\item{n_thread}{Number of parallel threads that will be created. This number
must be in accordance to the number of cores of the PC}

\item{swat_vers}{Character string that defines the SWAT version. Either
"2012" or "plus".}

\item{quiet}{Logical. Defines if messages should be written or function
should be executed quietly.}
}
\description{
Generate folder structure for parallel SWAT execution
}
\keyword{internal}
