% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{define_output}
\alias{define_output}
\title{Define output variables}
\usage{
define_output(file, variable = NULL, unit = NULL, expression = NULL)
}
\arguments{
\item{file}{The SWAT output output_file where the output variable is located (so
far only .sub and .rch are supported!)}

\item{variable}{The name of the variable as given in the header of the
respective output output_file (without unsits!)}

\item{unit}{The 'RCH' or 'SUB unit for which the output should be
written.}

\item{expr}{Caution! Advanced setting if the output variable should be
modified or aggregated before writing the output. To do so a piped
workflow must be provided (see examples).}
}
\value{
The function should only be used as a helper function for
  \code{define_output()}
}
\description{
Define which output variables (from the parallel SWAT execution of the
function \code{\link{run_swat}}) should be saved and returned after finishing
the model runs. Note that, the variable \code{output} in
\code{\link{run_swat}} must be defined via this function.
Helper function for the \code{\link{run_swat}}) output variable definition with
\code{\link{define_output}}
}
\details{
Each variable provided with the function \code{\link{define_output}}should be
defined using the helper function \code{\link{variable}}.
}
\examples{
# Writing the discharge for the RCH unsits 1 to 5
# and the nitrogen load and ET for the RCH or SUB unit 5:
out_tbl <- define_output(flow = out_var(output_file   = "rch",
                                        variable = "FLOW_OUT",
                                        spatial_unit  = 1:5),
                         no3  = out_var(output_file   = "rch",
                                        variable = "NO3_OUT",
                                        spatial_unit  = 5),
                         ptot = out_var(output_file   = "rch",
                                        variable = "PTOT",
                                        spatial_unit  = 5),
                         et_a = out_var(output_file   = "sub",
                                        variable = "ET",
                                        spatial_unit  = 5))


# Writing the dicharge for RCH 16,28, and 31 and the sum of discharge for
# all subbasins and time steps (as defined in sum_expr):

sum_expr <- "dplyr::filter(MON <= 366) \%>\%
               dplyr::select(FLOW_OUT) \%>\%
               dplyr::summarise(FLOW_SUM = sum(FLOW_OUT))"

out_tbl <- define_output(flow = out_var(output_file   = "rch",
                                        variable = "FLOW_OUT",
                                        spatial_unit  = c(16,28,31)),
                         flow_sum = out_var(output_file = "rch",
                                            expression  = sum_expr,
                                            add_date    = FALSE))

}

