% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{define_output}
\alias{define_output}
\title{Define output variables}
\usage{
define_output(file, variable = NULL, unit = NULL, expression = NULL)
}
\arguments{
\item{file}{Character string. The SWAT output output_file where the output
variable is located (Valid inputs are 'rch', 'sub', 'hru', and 'sed' for
the respective SWAT output files 'output.rch', 'output.sub', output.hru',
and 'output.sed')}

\item{variable}{Character string. Output variable available from the
respective SWAT output file. respective output file defined with
\code{file}. For the correct definition of the output variables please use
the variable names as defined in the
\href{https://swat.tamu.edu/media/69395/ch32_output.pdf}{SWAT Output Data
Documentation}. Optional, also the number of the respective column in the
output file can be provided.}

\item{unit}{Numeric vector. The spatial unit (e.g. the reach, subbasin, or
HRU) defined by the columns 'RCH', 'SUB', 'HRU' in the respective output
file for which the outputs should be extracted.}

\item{expr}{(optional) Alternative to \code{variable} and \code{unit}. Text
string of expression to apply to the output table to extract and modify a
variable before it is returned. To do so a piped workflow must be provided
(see examples).}
}
\description{
Define the SWAT2012 output variables that should be extracted after the SWAT
model execution and returned to R. Note that, the variable \code{output} in
\code{\link{run_swat2012}} must be defined via this function.
}
\examples{
# Writing the discharge for the RCH unsits 1 to 5
# and the nitrogen load and ET for the RCH or SUB unit 5:
out_tbl <- define_output(flow = out_var(output_file   = "rch",
                                        variable = "FLOW_OUT",
                                        spatial_unit  = 1:5),
                         no3  = out_var(output_file   = "rch",
                                        variable = "NO3_OUT",
                                        spatial_unit  = 5),
                         ptot = out_var(output_file   = "rch",
                                        variable = "PTOT",
                                        spatial_unit  = 5),
                         et_a = out_var(output_file   = "sub",
                                        variable = "ET",
                                        spatial_unit  = 5))


# Writing the dicharge for RCH 16,28, and 31 and the sum of discharge for
# all subbasins and time steps (as defined in sum_expr):

sum_expr <- "dplyr::filter(MON <= 366) \%>\%
               dplyr::select(FLOW_OUT) \%>\%
               dplyr::summarise(FLOW_SUM = sum(FLOW_OUT))"

out_tbl <- define_output(flow = out_var(output_file   = "rch",
                                        variable = "FLOW_OUT",
                                        spatial_unit  = c(16,28,31)),
                         flow_sum = out_var(output_file = "rch",
                                            expression  = sum_expr,
                                            add_date    = FALSE))

}

