% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_expression}
\alias{evaluate_expression}
\title{Evaluate the expression defined for a variable in 'output'}
\usage{
evaluate_expression(table, expr)
}
\arguments{
\item{table}{Table to which dplyr expression should be applied}

\item{expr}{Expression to be applied to table}
}
\description{
Evaluate the expression defined for a variable in 'output'
}
\keyword{internal}
