% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{get_value_range}
\alias{get_value_range}
\title{Get the value range from a condition in the par_def table}
\usage{
get_value_range(condition)
}
\arguments{
\item{condition}{Character string that defines a condition}
}
\description{
Get the value range from a condition in the par_def table
}
\keyword{internal}
