% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{initialize_save_file}
\alias{initialize_save_file}
\title{Initialize the data base wher model outputs are saved}
\usage{
initialize_save_file(save_path, parameter, model_setup)
}
\arguments{
\item{save_path}{Character string. Path of the sql data base}

\item{parameter}{Parameter set provided for simualtion}

\item{model_setup}{List with files and variables that define the SWAT model
setup}
}
\description{
Initialize the data base wher model outputs are saved
}
\keyword{internal}
