% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_functions.R
\name{initialize_save_file}
\alias{initialize_save_file}
\title{Initialize the data base wher model outputs are saved}
\usage{
initialize_save_file(save_path, parameter, run_info, run_index)
}
\arguments{
\item{save_path}{Path of the sql data base for incrementally saving simulations}

\item{parameter}{Parameter set provided for simulation}

\item{run_info}{List with meta data on simulation setup}

\item{run_index}{IDs of parameter sets for which simulations should be run}
}
\description{
Initialize the data base wher model outputs are saved
}
\keyword{internal}
