% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_handling.R
\name{load_demo}
\alias{load_demo}
\title{Load SWAT demo data}
\usage{
load_demo(dataset, path = NULL, swat_version = NULL)
}
\arguments{
\item{dataset}{Character string to define data set to load. Valid inputs are
\code{dataset = c('project', 'observation', 'subbasin', 'river', 'hru')}.
\code{dataset = 'project'} loads a SWAT demo project in the defined
\code{path} location and returns the final project path as a text string in
R. The definition of the \code{swat_version} is required.
\code{dataset = 'observation'} returns a \code{tibble} with discharge observation
data at the main outlet of the demo watershed.
The options \code{'subbasin'}, \code{'river'}, and \code{'hru'} return the paths
of the respective subbasin, river network and HRU shape files. See examples
how to load shapes unsing e.g. the \code{sf} package. These options require
a definition of the \code{swat_version}.}

\item{path}{Character string that defines the path where copy the SWAT demo
project.}

\item{swat_version}{Character string that defines the SWAT version. Options
are \code{swat_version = c('2012', 'plus')}. This argument is required to
load SWAT projects and shape files.}
}
\description{
Load SWAT demo data
}
\section{Examples}{

To learn how to load the different demo data sets with \code{SWATplusR} see the
section \href{https://chrisschuerz.github.io/SWATplusR/articles/SWATplusR.html#loading-demos}{Loading demos}
on the package's \emph{Get Started} page.
}

