% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_parameter_modify.R
\name{modify_parameter}
\alias{modify_parameter}
\title{Modify the model parameters of the thread that runs the simulation run_i}
\usage{
modify_parameter(parameter, thread_parameter, file_meta, run_index, i_run)
}
\arguments{
\item{parameter}{List providing the parameter table and the parameter
constraints}

\item{thread_parameter}{The swat_parameter set that will be modified
according to the parpameter set used in the respective thread}

\item{file_meta}{Table with the meta informations on the SWAT model parameter
files}

\item{run_index}{Vector of the indices of runs that are performed}

\item{i_run}{Index that gives the number of the current run simulated in the
respective thread}
}
\description{
Modify the model parameters of the thread that runs the simulation run_i
}
\keyword{internal}
