% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_output.R
\name{prepare_output_definition}
\alias{prepare_output_definition}
\title{Check output if is a data.frame and convert in case to named list}
\usage{
prepare_output_definition(output, swat_vers, project_path)
}
\arguments{
\item{output}{Output (or list of outputs) defined with \code{define_output()}}

\item{swat_vers}{SWAT version one of '2012' or 'plus'}
}
\description{
Check output if is a data.frame and convert in case to named list
}
\keyword{internal}
