% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_handle_output.R
\name{read_mgtout}
\alias{read_mgtout}
\title{Read and process SWAT+ management outputs.}
\usage{
read_mgtout(output_i, thread_path)
}
\arguments{
\item{output_i}{i_th part from the output table which defines what to}

\item{thread_path}{String path to the thread where to read the output file}
}
\description{
Read and process SWAT+ management outputs.
}
\keyword{internal}
