% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a set of scenarios for a SWAT+/SWAT2012 project}
\usage{
run_scenario(
  project_path,
  scenario_path,
  output,
  version,
  n_thread = NULL,
  ...
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{output}{Output variables to extract and return for each SWAT model run.

Each output variable which should be returned is defined with \code{\link{define_output}}.
If multiple output variables should be returned for a model run, the output
definitions must be concatenated in a named list. See the examples in the
section below and on the help page of \code{\link{define_output}}.}

\item{n_thread}{(optional) Number parallel threads in which the simulations
are performed. Default \code{n_thread} is \code{NULL} and all simulations are run in
the folder '.model_run/thread_1' and on a single core. If \code{n_thread} is
defined with a number greater than 1 multiple thread folders are generated
and the simulations are performed in parallel on that number of cores.
\code{n_thread} is always limited by the number of cores of the used computer.}

\item{parameter}{(optional) Model parameter changes to be implemented in the
simulation runs.

Parameter changes can be defined by providing a named vector with the
changes for a single parameter combination or a tibble where each row is
a parameter combinantion which should be used in a model simulation.

For examples of the implementation of parameter changes in model simulations
see the
\href{https://chrisschuerz.github.io/SWATrunR/articles/SWATrunR.html#changing-parameter-values}{\code{SWATrunR}}
github page and the examples in corresponding section below.}

\item{start_date, end_date}{(optional) Start and end dates of the SWAT simulation.

Start and end dates have to be defined together. Default values are
\code{NULL}. In this case the start and end dates which are defined in file
'time.sim' of the SWAT+ project are used in the simulation runs.

The dates can be provided as character strings in any ymd format
(e.g. 'yyyy-mm-dd'), as numeric values (yyyymmdd) or in a \code{Date}
format.}

\item{years_skip, start_date_print}{(optional) Start year/date for output printing.

\code{years_skip} or \code{start_date_print} can be used to define the start year/date
in which the first simulation outputs should be written. The two arguments
are useful to define a 'warm-up' period for the simulation runs for which
the simulation results are not considered.

\code{years_skip} is provided as an integer value which defines the number of
years to skip before writing outputs.

\code{start_date_print} can be provided as character string in any ymd format
(e.g. 'yyyy-mm-dd'), as a numeric value (yyyymmdd) or in a \code{Date}
format.

For the simulation of hydrological years (which do not start at Jan 1.)
the use of \code{start_date_print} is the preferred option to define the actual
start date of the hydrological year (e.g. '2000-10-01').

\code{start_date_print} is always overruled by the definition of \code{years_skip}.

The default values of \code{years_skip} and \code{start_date_print} are \code{NULL}.
If both are \code{NULL} the projects' default setting for output printing which
is defined in the input file 'print.prt' is used.}

\item{run_index}{(optional) Vector of IDs of parameter combinations for which
simulations should be performed.

\code{run_index} is useful to only perform a subset of simulations for a defined
parameter set (e.g. when distributing simulations between several computers,
or when rerunning failed simulations for certain parameter combinations).

\code{run_index} is defined as a numeric vector where the values correspond to
the row IDs of the parameter input table. (e.g.\code{run_index = c(1:100,
  110, 115)}). The default value is \code{NULL}. Then the simulations for all
provided parameter combinations are performed.}

\item{run_path}{(optional) Path to a location where the simulations should be
performed.

\code{run_path} is useful if the project folder (in \code{project_path}) is in a
location where the simulations should not be performed (e.g. because of
limited space, or a virtual or slow drive). In such a case it can help to
perform the simulations on a different hard drive (e.g. local drive).

In the path which is defined with \code{run_path} the folder '.model_run' is
generated and the SWAT project is copied there into a number of sub-folders
'thread_\if{html}{\out{<i>}}', depending on the number of parallel runs which were defined.

With the default value of \code{run_path = NULL} the folder '.model_run' is
generated in the \code{project_path}.}

\item{save_file}{(optional) Name of the folder where simulation runs should be
saved incrementally.

If a \code{save_file} is provided a folder with that name is generated and the
simulation results are incrementally saved in several SQLite data bases in
the folder. This option is useful when very long simulation experiments are
performed and computer/simulation crashes may result in the loss of days or
weeks of simulation runs, or when simulation outputs are returned which
would not fit into the computers RAM.

Default is \code{NULL} and simulations are not saved incrementally.}

\item{save_path}{(optional) Path where the \code{save_file} folder should be
generated.

In the default case \code{save_path = NULL} the \code{save_file} folder is generated
in the \code{project_path}. If a \code{save_path} is provided the \code{save_file} folder
is generated there if \code{save_file} in not \code{NULL}.}

\item{return_output}{(optional) Return outputs in \code{R}. If \code{TRUE} (default
value) then the defined simulated output variables are returned at the end
of all simulation runs. If \code{FALSE} no simulations are returned in \code{R}.

\strong{Caution:} Please only set \code{return_output = FALSE} if the simulations are
saved in a \code{save_file}. Otherwise the simulations are performed but no
simulations are returned.}

\item{add_parameter}{(optional)  Add the used parameter changes in the
simulation outputs. If \code{TRUE} (default value and recommended) the returned
list with the simulation outputs includes the list element \code{.$parameter}
which provides the definition of the parameters \code{.$parameter$definition}
and the values of the parameter changes \code{.$parameter$values}.}

\item{add_date}{(optional) Add a date vector to simulated time series. If
\code{TRUE} (default value) a date vector is added.}

\item{split_units}{(optional) Split simulated outputs of the same variable into
separate tables in the \code{.$simulation} list. If \code{TRUE} (default value) outputs
are split (e.g. for \code{define_output('channel_sd_day', 'flo_out', 1:3)}) the
output tables \code{flo_out_1}, \code{flo_out_2}, and \code{flo_out_3} will be generated.
If \code{FALSE} in the same example a single table is generated with an integer
column \code{unit} with the ID values 1 to 3.}

\item{refresh}{(optional) Rewrite existing '.model_run' folder. If \code{TRUE}
(default value and recommended) always forces that .model_run' is newly
written when SWAT run ins started.}

\item{keep_folder}{(optional) Keep the thread folders in '.model_run' after
finishing the simulation runs. If \code{FALSE} (default value) the '.model_run'
folder is deleted after the last simulation. Keeping the thread folders
(\code{keep_folder = TRUE}) can be useful to manually check input and output
files in the thread folders after the simulation.}

\item{quiet}{(optional) Printing the progress of the simulations. If
\code{TRUE} (default value) the progress of the simulation runs is printed in the
console.}

\item{time_out}{(optional) Timeout for simulation runs in seconds.

Simulations may hang up due to specific parameter combinations. A timeout
kills any simulation if the run time exceeds the set time in seconds.
Be careful with this setting. A timeout which is set shorter than the
actual run time of a simulation will also kill all potentially successful
runs before finishing.

Default \code{time_out} is set to \code{Inf} to not kill any simulations.}
}
\value{
Returns the simulation results as list with the following elements:
\itemize{
\item \code{.$parameter} is only available if parameter changes were implemented in
the simulation runs. \code{.$parameter} is a list with 2 elements:
\itemize{
\item \code{.$parameter$definition} is a tibble which shows how the parameter
changes were defined.
\item \code{.$parameter$values} is a tibble with the values of the defined
parameter changes. Each row is a parameter set. The names of the
columns correspond to the \code{par_name} which are defined in
\code{.$parameter$definition}.
}
\item \code{.$simulation} is a list with the defined and simulated output variables.
Each list element is one defined output variable. The name of each list
element is the name which was defined with \code{define_output()}.
\item \code{.$error_report} is only available if simulation runs failed. This element
is a tibble which summarizes the failed simulations and the triggered errors
of the SWAT executable which caused a simulation to fail.
\item \code{.$run_info} is a list with meta information on the simulation run. The
saved information is:
\itemize{
\item \code{.$run_info$simulation_log} is a tibble which logs simulation start and
end time_stamps, duration and paths of the project folder.
\item \code{.$run_info$simulation_period} is a tibble which provides dates, and skipped
years of the saved output variables.
\item \code{.$run_info$output_definition} is a tibble which summarizes the defined
output variables which were passed with the input argument \code{output} and
defined with \code{define_output()}.
}
}
}
\description{
\code{run_swatplus()} performs simulations for a SWAT+ model setup that is located
in the folder \code{project_path} and returns simulated output variables which are
defined with \code{output} in a tidy format. Optionally, a set of parameter changes
can be defined with \code{parameter} which are then applied in the simulation runs.

Several settings such as the definition of the simulation period, incremental
saving of simulation results into SQLite data bases, or the use of multiple
cores for the simulation runs can be defined.
}
\examples{
# Install the SWATdata R package which provides a SWAT+ demo project
if(!'SWATdata' \%in\% installed.packages()) {
  remotes::install_github('chrisschuerz/SWATdata')
}

# Use a temporary dir for the demo project
tmp_dir <- tempdir()

# Load a SWAT+ demo project
proj_path <- load_demo(dataset = 'project',
                       path = tmp_dir,
                       version = 'plus')

# Perform simulations for the demo project and return
# daily simulated discharge for channel 1
q_sim <- run_swatplus(project_path = proj_path,
                      output = define_output(file = 'channel_sd_day',
                                             variable = 'flo_out',
                                             unit = 1))

# The simulated time series in a tibble format
q_sim$simulation$flo_out

# Plot the simulated time series
plot(q_sim$simulation$flo_out, type = 'l')

# Print the meta information for the simulation run
q_sim$run_info

# Delete demo project folder
unlink(tmp_dir, recursive = TRUE, force = TRUE)

}
