% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a set of scenarios for a SWAT+/SWAT2012 project}
\usage{
run_scenario(
  project_path,
  scenario_path,
  output,
  version,
  n_thread = NULL,
  ...
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{scenario_path}{Path to the folder on the hard drive in which the
scenarios are organized. The scenarios must be defined in sub-folders. All
files of a sub-folder will overwrite the respective files of the SWAT
project. Differences between the original files and the scenario files
define a scenario. The scenario outputs will have the same name as the
respective scenario folders.}

\item{output}{Output variables to extract and return for each SWAT model run.

Each output variable which should be returned is defined with
\code{\link{define_output}}.
If multiple output variables should be returned for a model run, the output
definitions must be concatenated in a named list. See the examples provided
on the help page of the functions \code{\link{run_swatplus}} and
\code{\link{run_swat2012}} and on the help page of
\code{\link{define_output}}.}

\item{version}{The version of the SWAT model setup in \code{project_path}.
\code{version} must be either \code{'plus'} for SWAT+ model setups and \code{'2012'} if
the model setup is a SWAT2012 project.}

\item{n_thread}{(optional) Number parallel threads in which the simulations
are performed. Default, \code{n_thread} is \code{NULL} and all simulations are run on
a single core. If \code{n_thread} is defined with a number greater than 1 the
simulations are performed in parallel on that number of cores.
\code{n_thread} is always limited by the number of cores of the used computer.}

\item{...}{Other input arguments passed on to \code{run_swatplus()} or
\code{run_swat2012()} depending on the \code{version} of the SWAT project. All input
arguments passed with \code{...} must be defined for the respective run function.}
}
\description{
\code{run_scenario()} uses scenario input files which are organized in the
\code{scenario_path} and performs scenario simulations for a SWAT+/SWAT2012 model
setup, which is located in the \code{project_path}. The scenario simulations
return output variables which are defined with \code{output}. The version of the
model set up is defined with \code{version}. Scenarios can be run in parallel by
defining \code{n_thread}.

Further input arguments for \code{run_swatplus()} and \code{run_swat2012()} can be
passed to \code{run_scenario()}. E.g. if a parameter set is provided with
\code{parameter} all defined parameter combinations will be run for each scenario.
}
