% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat.R
\name{run_swat2012}
\alias{run_swat2012}
\title{Run a SWAT project and return ouputs in R}
\usage{
run_swat2012(project_path, output, parameter = NULL, start_date = NULL,
  end_date = NULL, output_interval = NULL, years_skip = NULL,
  rch_out_var = NULL, sub_out_var = NULL, hru_out_var = NULL,
  hru_out_nr = NULL, abs_swat_val = NULL, run_index = NULL,
  run_path = NULL, n_thread = NULL, save_file = NULL, save_incr = FALSE,
  save_parameter = TRUE, return_out = TRUE, refresh = FALSE,
  keep_folder = FALSE, quiet = FALSE)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{output}{Define the output variables to extract from the SWAT model
runs. See function \code{\link{define_output}} help file to see how to
define an output.}

\item{parameter}{(optional) SWAT model parameters either provided as named
vector or data.frame. If \code{parameter} is provided respective parameters
are modified accordingly.}

\item{start_date}{(ptional) Start date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format
project are located}

\item{end_date}{(optional) End date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format
project are located}

\item{output_interval}{(optional) Time interval in which the SWAT model
outputs are written. Provided either as character string ("d" for daily,
"m" for monthly, or "y" for yearly) or as SWAT input values (0 for monthly,
1 for daily, 2 for yearly).}

\item{years_skip}{(optional) Integer value that provides the numbe of years
to be skipped during writing the SWAT model outputs}

\item{rch_out_var}{(optional) Numeric vector of maximum length = 20 for
customized output of reach variables.For output codes see
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.77ff.}

\item{sub_out_var}{(optional) Numeric vector of maximum length = 15 for
customized output of subbasin variables.For output codes see
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.78ff.}

\item{hru_out_var}{(optional) Numeric vector of maximum length = 20 for
customized output of HRU variables.For output codes see
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.79ff.}

\item{hru_out_nr}{(optional) Numeric vector of maximum length = 20 for
providing the HRU numbers for which the HRU variables are written. Optional
if hru_out_nr = 'all', HRU variables are written for all HRU (caution, very
large output files possible!)}

\item{abs_swat_val}{(optional) \code{run_swat2012} uses an internal
'Absolute_SWAT_Values.txt' file required for overwriting parameters. With
this parameter the path to a custom 'Absolute_SWAT_Values' file can be
provided in case it is necessary.}

\item{run_index}{(optional) Numeric vector (e.g.\code{run_index = c(1:100,
110, 115)}) to run a subset of the provided \code{parameter} sets. If NULL
all provided parameter sets are used.}

\item{run_path}{(optional) Character string that provides the path where the
'.model_run' folder is written and the SWAT models are executed. If NULL
'.model_run' is built in the project folder.}

\item{n_thread}{(optional) Number of threads to be used for the parallel
model run. If not provided models are run on single core}

\item{save_file}{(optional) Character string that provides file path to which
model results are saved after all model runs are completed. (Load
\code{save_file} with \code{file <- readRDS(save_file)}).}

\item{save_incr}{(optional) Logical. If \code{save_incr = TRUE} model runs
are saved incrementally and can be restored with \code{\link{recover_run}}.}

\item{save_parameter}{(optional) Logical. If \code{save_parameter = TRUE}
used parameter sets are saved and/or returned together with the model
outputs.}

\item{return_out}{(optional) Logical. Whether outputs should be returned or
not. Set \code{return_out = FALSE} and provide \code{save_file} if outputs
should only be saved on hard drive.}

\item{refresh}{(optional) Logical. \code{refresh = TRUE} always forces that
'.model_run' is newly written when SWAT run ins started.}

\item{keep_folder}{(optional) Logical. If \code{keep_folder = TRUE}
'.model_run' is kept and not deleted after finishing model runs. In this
case '.model_run' is reused in a new model run if \code{refresh = FALSE}.}

\item{quiet}{(optional) Logical. If \code{quiet = TRUE} no messages are
written.}
}
\value{
Returns the simulation results for the defined output variables as a
  named list. If more than one parameter set was provided the list contains a
  tibble where each column is a model run.
}
\description{
Run a SWAT project and return ouputs in R
}

