% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat2012_run_swat.R
\name{run_swat2012}
\alias{run_swat2012}
\title{Run simulations for a SWAT2012 project}
\usage{
run_swat2012(
  project_path,
  output,
  parameter = NULL,
  start_date = NULL,
  end_date = NULL,
  output_interval = NULL,
  years_skip = NULL,
  rch_out_var = NULL,
  sub_out_var = NULL,
  hru_out_var = NULL,
  hru_out_nr = NULL,
  run_index = NULL,
  run_path = NULL,
  n_thread = NULL,
  save_file = NULL,
  save_path = NULL,
  return_output = TRUE,
  add_parameter = TRUE,
  add_date = TRUE,
  split_units = FALSE,
  refresh = TRUE,
  keep_folder = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{output}{Output variables to extract and return for each SWAT model run.

Each output variable which should be returned is defined with \code{\link{define_output}}.
If multiple output variables should be returned for a model run, the output
definitions must be concatenated in a named list. See the examples in the
section below and on the help page of \code{\link{define_output}}.}

\item{parameter}{(optional) Model parameter changes to be implemented in the
simulation runs.

Parameter changes can be defined by providing a named vector with the
changes for a single parameter combination or a tibble where each row is
a parameter combinantion which should be used in a model simulation.

For examples of the implementation of parameter changes in model simulations
see the
\href{https://chrisschuerz.github.io/SWATplusR/articles/SWATrunR.html}{\code{SWATrunR}}
github page and the examples in corresponding section below.}

\item{start_date, end_date}{(optional) Start and end dates of the SWAT simulation.

Start and end dates have to be defined together. Default values are
\code{NULL}. In this case the start and end dates which are defined in file
'time.sim' of the SWAT+ project are used in the simulation runs.

The dates can be provided as character strings in any ymd format
(e.g. 'yyyy-mm-dd'), as numeric values (yyyymmdd) or in a \code{Date}
format.}

\item{output_interval}{(optional) Time interval in which the SWAT model
outputs are written.

If \code{NULL} (default) the value which is defined in the file.cio is used.
\code{output_interval} can be defined either with a character string where \code{'d'}
would result in daily output, and \code{'m'} in monthly, and \code{'y'} in yearly
outputs. Alternatively also the as SWAT2012 input values (0 for monthly,
1 for daily, 2 for yearly) are allowed.}

\item{years_skip}{(optional) Number of years to skip before printing outputs.

\code{years_skip} can be used to define the start year in which the first
simulation outputs should be written. Hence, \code{years_skip} is useful to
define a 'warm-up' period for the simulation runs for which the simulation
results are not considered.

\code{years_skip} is provided as an integer value which defines the number of
years to skip before writing outputs.

The default values of \code{years_skip} is \code{NULL}. In this case the projects'
default setting for output printing is used which is defined in the input
file file.cio'.}

\item{rch_out_var}{(optional) Numeric vector of maximum \code{length = 20} for
customized output of reach variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.77ff.}

\item{sub_out_var}{(optional) Numeric vector of maximum \code{length = 15} for
customized output of subbasin variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.78ff.}

\item{hru_out_var}{(optional) Numeric vector of maximum \code{length = 20} for
customized output of HRU variables.

For output codes see the
\href{https://swat.tamu.edu/media/69308/ch03_input_cio.pdf}{SWAT I/O
Documentation} p.79ff.}

\item{hru_out_nr}{(optional) Numeric vector of maximum \code{length = 20} for
providing the HRU numbers for which the HRU variables are written.

Optional if \code{hru_out_nr = 'all'}, HRU variables are written for all HRU
(caution, very large output files possible!)}

\item{run_index}{(optional) Vector of IDs of parameter combinations for which
simulations should be performed.

\code{run_index} is useful to only perform a subset of simulations for a defined
parameter set (e.g. when distributing simulations between several computers,
or when rerunning failed simulations for certain parameter combinations).

\code{run_index} is defined as a numeric vector where the values correspond to
the row IDs of the parameter input table. (e.g.\code{run_index = c(1:100,
  110, 115)}). The default value is \code{NULL}. Then the simulations for all
provided parameter combinations are performed.}

\item{run_path}{(optional) Path to a location where the simulations should be
performed.

\code{run_path} is useful if the project folder (in \code{project_path}) is in a
location where the simulations should not be performed (e.g. because of
limited space, or a virtual or slow drive). In such a case it can help to
perform the simulations on a different hard drive (e.g. local drive).

In the path which is defined with \code{run_path} the folder '.model_run' is
generated and the SWAT project is copied there into a number of sub-folders
'thread_\if{html}{\out{<i>}}', depending on the number of parallel runs which were defined.

With the default value of \code{run_path = NULL} the folder '.model_run' is
generated in the \code{project_path}.}

\item{n_thread}{(optional) Number parallel threads in which the simulations
are performed. Default \code{n_thread} is \code{NULL} and all simulations are run in
the folder '.model_run/thread_1' and on a single core. If \code{n_thread} is
defined with a number greater than 1 multiple thread folders are generated
and the simulations are performed in parallel on that number of cores.
\code{n_thread} is always limited by the number of cores of the used computer.}

\item{save_file}{(optional) Name of the folder where simulation runs should be
saved incrementally.

If a \code{save_file} is provided a folder with that name is generated and the
simulation results are incrementally saved in several SQLite data bases in
the folder. This option is useful when very long simulation experiments are
performed and computer/simulation crashes may result in the loss of days or
weeks of simulation runs, or when simulation outputs are returned which
would not fit into the computers RAM.

Default is \code{NULL} and simulations are not saved incrementally.}

\item{save_path}{(optional) Path where the \code{save_file} folder should be
generated.

In the default case \code{save_path = NULL} the \code{save_file} folder is generated
in the \code{project_path}. If a \code{save_path} is provided the \code{save_file} folder
is generated there if \code{save_file} in not \code{NULL}.}

\item{return_output}{(optional) Return outputs in \code{R}. If \code{TRUE} (default
value) then the defined simulated output variables are returned at the end
of all simulation runs. If \code{FALSE} no simulations are returned in \code{R}.

\strong{Caution:} Please only set \code{return_output = FALSE} if the simulations are
saved in a \code{save_file}. Otherwise the simulations are performed but no
simulations are returned.}

\item{add_parameter}{(optional)  Add the used parameter changes in the
simulation outputs. If \code{TRUE} (default value and recommended) the returned
list with the simulation outputs includes the list element \code{.$parameter}
which provides the definition of the parameters \code{.$parameter$definition}
and the values of the parameter changes \code{.$parameter$values}}

\item{add_date}{(optional) Add a date vector to simulated time series. If
\code{TRUE} (default value) a date vector is added.}

\item{refresh}{(optional) Rewrite existing '.model_run' folder. If \code{TRUE}
(default value and recommended) always forces that .model_run' is newly
written when SWAT run ins started.}

\item{keep_folder}{(optional) Keep the thread folders in '.model_run' after
finishing the simulation runs. If \code{FALSE} (default value) the '.model_run'
folder is deleted after the last simulation. Keeping the thread folders
(\code{keep_folder = TRUE}) can be useful to manually check input and output
files in the thread folders after the simulation.}

\item{quiet}{(optional) Printing the progress of the simulations. If
\code{TRUE} (default value) the progress of the simulation runs is printed in the
console.}
}
\value{
Returns the simulation results as list with the following elements:
\itemize{
\item \code{.$parameter} is only available if parameter changes were implemented in
the simulation runs. \code{.$parameter} is a list with 2 elements:
\itemize{
\item \code{.$parameter$definition} is a tibble which shows how the parameter
changes were defined.
\item \code{.$parameter$values} is a tibble with the values of the defined
parameter changes. Each row is a parameter set. The names of the
columns correspond to the \code{par_name} which are defined in
\code{.$parameter$definition}.
}
\item \code{.$simulation} is a list with the defined and simulated output variables.
Each list element is one defined output variable. The name of each list
element is the name which was defined with \code{define_output()}.
\item \code{.$error_report} is only available if simulation runs failed. This element
is a tibble which summarizes the failed simulations and the triggered errors
of the SWAT executable which caused a simulation to fail.
\item \code{.$run_info} is a list with meta information on the simulation run. The
saved information is:
\itemize{
\item \code{.$run_info$simulation_log} is a tibble which logs simulation start and
end time_stamps, duration and paths of the project folder.
\item \code{.$run_info$simulation_period} is a tibble which provides dates, and skipped
years of the saved output variables.
\item \code{.$run_info$output_definition} is a tibble which summarizes the defined
output variables which were passed with the input argument \code{output} and
defined with \code{define_output()}.
}
}
}
\description{
\code{run_swat2012()} performs simulations for a SWAT2012 model setup that is located
in the folder \code{project_path} and returns simulated output variables which are
defined with \code{output} in a tidy format. Optionally, a set of parameter changes
can be defined with \code{parameter} which are then applied in the simulation runs.

Several settings such as the definition of the simulation period, incremental
saving of simulation results into SQLite data bases, or the use of multiple
cores for the simulation runs can be defined.
}
\examples{
# Install the SWATdata R package which provides a SWAT+ demo project
if(!'SWATdata' \%in\% installed.packages()) {
  remotes::install_github('chrisschuerz/SWATdata')
}

# Use a temporary dir for the demo project
tmp_dir <- tempdir()

# Load a SWAT2012 demo project
proj_path <- load_demo(dataset = 'project',
                       path = tmp_dir,
                       version = '2012')

# Perform simulations for the demo project and return
# daily simulated discharge for channel 1
q_sim <- run_swat2012(project_path = proj_path,
                      output = define_output(file = 'rch',
                                             variable = 'FLOW_OUT',
                                             unit = 1))

# The simulated time series in a tibble format
q_sim$simulation$FLOW_OUT

# Plot the simulated time series
plot(q_sim$simulation$FLOW_OUT, type = 'l')

# Print the meta information for the simulation run
q_sim$run_info

# Delete demo project folder
unlink(tmp_dir, recursive = TRUE, force = TRUE)


}
