% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_run_swat.R
\name{run_swatplus}
\alias{run_swatplus}
\title{Run SWAT+}
\usage{
run_swatplus(project_path, output, parameter = NULL, start_date = NULL,
  end_date = NULL, output_interval = NULL, years_skip = NULL,
  run_index = NULL, run_path = NULL, n_thread = NULL,
  save_path = NULL, save_file = NULL, return_output = TRUE,
  add_parameter = TRUE, add_date = TRUE, refresh = TRUE,
  keep_folder = FALSE, quiet = FALSE)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{output}{Define the output variables to extract from the SWAT model
runs. See function \code{\link{define_output}} help file to see how to
define an output.}

\item{parameter}{(optional) SWAT model parameters either provided as named
vector or data.frame. If \code{parameter} is provided respective parameters
are modified accordingly.}

\item{start_date}{(optional) Start date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format
project are located.}

\item{end_date}{(optional) End date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format
project are located}

\item{output_interval}{(optional) Time interval in which the SWAT model
outputs are written. Provided either as character string ("d" for daily,
"m" for monthly, or "y" for yearly) or as SWAT input values (0 for monthly,
1 for daily, 2 for yearly).}

\item{years_skip}{(optional) Integer value that provides the numbe of years
to be skipped during writing the SWAT model outputs}

\item{run_index}{(optional) Numeric vector (e.g.\code{run_index = c(1:100,
110, 115)}) to run a subset of the provided \code{parameter} sets. If NULL
all provided parameter sets are used.}

\item{run_path}{(optional) Character string that provides the path where the
'.model_run' folder is written and the SWAT models are executed. If NULL
'.model_run' is built in the project folder.}

\item{n_thread}{(optional) Number of threads to be used for the parallel
model run. If not provided models are run on single core}

\item{save_path}{(optional) Character string to define the path where the
model runs are saved if save_file is defined. If \code{save_path = NULL}
the \code{save_file} is saved in the project_path.}

\item{save_file}{(optional) Character string to define the name of the file
where the simulations are saved.}

\item{return_output}{(optional) Logical. Whether outputs should be returned
or not. Set \code{return_out = FALSE} and provide \code{save_file} if
outputs should only be saved on hard drive.  \code{Default = TRUE}}

\item{add_parameter}{(optional) Logical. If \code{add_parameter = TRUE}
used parameter sets are saved and/or returned together with the model
outputs.  \code{Default = TRUE}}

\item{add_date}{(optional) Logical. If \code{add_date = TRUE} a date column
is added to every simulatiuon output table.  \code{Default = TRUE}}

\item{refresh}{(optional) Logical. \code{refresh = TRUE} always forces that
'.model_run' is newly written when SWAT run ins started. \code{Default =
TRUE}}

\item{keep_folder}{(optional) Logical. If \code{keep_folder = TRUE}
'.model_run' is kept and not deleted after finishing model runs. In this
case '.model_run' is reused in a new model run if \code{refresh = FALSE}.
\code{Default = FALSE}}

\item{quiet}{(optional) Logical. If \code{quiet = TRUE} no messages are
written.  \code{Default = FALSE}}
}
\value{
Returns the simulation results for the defined output variables as a
  tibble. If more than one parameter set was provided the list of tibbles is
  returned where each column is a model run and each list entry is an output
  variable.
}
\description{
This function allows to run a SWAT+ project in parallel from within R.
Basic settings for the SWAT run such as the simulation period or the time
interval for the outputs can be done directly. SWAT simulation outputs can be
defined that are returned for defined parameter sets.
}
