% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{setup_calibration_cal}
\alias{setup_calibration_cal}
\title{Setup the calibration.cal file and include all parameter conditions
of the simulation and set them according to respective input parameters}
\usage{
setup_calibration_cal(par_def, unit_cons)
}
\arguments{
\item{par_def}{Tibble with one line that includes the parameter definition of
parameter i}

\item{unit_cons}{List of tibbles that contains the meta information of hru, aqu,
etc. units and constrain variables (e.g. texture, plant)}
}
\description{
Setup the calibration.cal file and include all parameter conditions
of the simulation and set them according to respective input parameters
}
\keyword{internal}
