% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{setup_swatplus}
\alias{setup_swatplus}
\title{Reads and modifies the SWAT+ projects' files that define the basic settings
of the simulation and set them according tu respective input parameters}
\usage{
setup_swatplus(
  project_path,
  parameter,
  output,
  start_date,
  end_date,
  start_date_print,
  years_skip,
  unit_cons
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{start_date}{Start date of the SWAT simulation. Provided as character
string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format project are
located}

\item{end_date}{End date of the SWAT simulation. Provided as character string
in any ymd format (e.g. 'yyyy-mm-dd') or in Date format project are located}

\item{years_skip}{Integer value that provides the numbe of years to be
skipped during writing the SWAT model outputs}
}
\description{
Reads and modifies the SWAT+ projects' files that define the basic settings
of the simulation and set them according tu respective input parameters
}
\keyword{internal}
