% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_outputs.R
\name{tidy_results}
\alias{tidy_results}
\title{Tidy up simulation results before returning them}
\usage{
tidy_results(sim_result, parameter, date, add_parameter, add_date, run)
}
\arguments{
\item{parameter}{Provided parameter set}

\item{add_parameter}{Logical. If TRUE parameters are saved in outputs}

\item{add_date}{Logical. If TRUE Dates are added to the simulation results}

\item{sim_results}{Extracted simulation results from the SWAT model runs}

\item{file_cio}{Modified file.cio}
}
\description{
Tidy up simulation results before returning them
}
\keyword{internal}
