% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_setup_run_files.R
\name{write_swatplus_setup}
\alias{write_swatplus_setup}
\title{Write the updated init files to all parallel folders}
\usage{
write_swatplus_setup(run_path, model_setup)
}
\arguments{
\item{run_path}{Path to the .model_run folder}

\item{model_setup}{List of files that define the SWAT+ model setup}
}
\description{
Write the updated init files to all parallel folders
}
\keyword{internal}
